+++ inc/progmesh.cpp
@@ -21,8 +21,8 @@
 
 namespace ProgMesh {
 
-template<class T> int   Contains(const std::vector<T> & c, const T & t){ return std::count(begin(c), end(c), t); }
-template<class T> int   IndexOf(const std::vector<T> & c, const T & v) { return std::find(begin(c), end(c), v) - begin(c); }
+template<class T> int   Contains(const std::vector<T> & c, const T & t){ return (int)std::count(begin(c), end(c), t); }
+template<class T> int   IndexOf(const std::vector<T> & c, const T & v) { return (int)(std::find(begin(c), end(c), v) - begin(c)); }
 //template<class T> T &   Add(std::vector<T> & c, T t)                   { c.push_back(t); return c.back(); }
 template<class T> T     Pop(std::vector<T> & c)                        { const T val = std::move(c.back()); c.pop_back(); return val; }
 template<class T> void  AddUnique(std::vector<T> & c, T t)             { if (!Contains(c, t)) c.push_back(t); }
@@ -157,7 +157,7 @@ inline void Triangle::ReplaceVertex(Vertex * const vold, Vertex * const vnew)
 __forceinline Vertex::Vertex(const float3 &v, const size_t _id)
 {
 	position = v;
-	id = _id;
+	id = (unsigned int)_id;
 	vertices.push_back(this);
 }
 
@@ -344,7 +344,7 @@ void ProgressiveMesh(const std::vector<float3> &vert, const std::vector<tridata>
 		// get the next vertex to collapse
 		Vertex *mn = MinimumCostEdge();
 		// keep track of this vertex, i.e. the collapse ordering
-		permutation[mn->id] = vertices.size() - 1;
+		permutation[mn->id] = (unsigned int)(vertices.size() - 1);
 		// keep track of vertex to which we collapse to
 		map[vertices.size() - 1] = mn->collapse ? mn->collapse->id : ~0u;
 		// Collapse this edge
